+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
+## [unreleased]
+### Added
+- Hovertip to gene panel names with associated genes in variant view, when variant covers more than one gene
+- Tests for panel to genes
+### Changed
+- Allow use of projections when retrieving gene panels
+
 ## [4.74.1]
 ### Changed
 - Parse and save into database also OMIM terms not associated to genes
@@ -14,15 +21,12 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [4.74]
 ### Added
-- Hovertip to gene panel names with associated genes in variant view, when variant covers more than one gene 
-- Tests for panel to genes
 - SNVs and Indels, MEI and str variants genes have links to Decipher
 - An `owner + case display name` index for cases database collection
 - Test and fixtures for RNA fusion case page
 - Load and display fusion variants from VCF files as the other variant types
 - Option to update case document with path to mei variants (clinical and research)
 ### Changed
-- Allow use of projections when retrieving gene panels
 - Details on variant type and category for audit filters on case general report
 - Enable Gens CN profile button also in somatic case view
 - Fix case of analysis type check for Gens analysis button - only show for WGS