+++ CHANGELOG.md
@@ -10,6 +10,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Tests for panel to genes
 ### Changed
 - Allow use of projections when retrieving gene panels
+- Do not save custom images as binary data into case and variant database documents
+- Retrieve and display case and variant custom images using image's saved path
 
 ## [4.74.1]
 ### Changed